+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: Added flexible service key matching with URL templating.
 - changed: Migrated to JSON-based logs using Pino library.
+- changed: Use serviceKeys for nownodes API key.
 - removed: Removed plugins to reduce load: abstract, amoy, arbitrum, avalanche, base, bobevm, celo, ethereumclassic, ethereumpow, fantom, filecoinfevm, filecoinfevmcalibration, holesky, hyperevm, pulsechain, rsk, sepolia, sonic.
 
 ## 0.2.6 (2025-12-10)